Adjusted configuration file structure.
Due to what seems to be a bug in `serde_yaml`, the configuration file
structure in 0.1.1 didn't allow for the use of OSC parameters that
required a tag to specify (i.e. any constant value).

Unfortunately, I wasn't able to find a fix that retained backwards
compatibility with the prior configuration file format without
maintaining a separate "backwards-compatibility" config file format, but
the changes needed are minor and the version is still early enough that
I'm not going to go through the effort to retain backwards-
compatibility.

0.2.0 it is!

While I'm at it, I also added flags to error on any unknown fields in
the config file, and made `on-change` an explicit section, rather than
implied, for forwards-compatibility reasons.