Sendr VMC output, OSC input, and TOML config.
This update adds VMC (technically still OSC, but more specialized) as a
configurable output type. This allows the Sender to act as an in-between
between a back-end motion capture software (e.g. VSeeFace) and a
front-end display (e.g. VNyan) to allow joystick inputs to affect the
avatar's pose, as well as setting blendshapes to allow for control over
props.

This also adds OSC as a potential *input* source. Due to the large
amount of data transferred, as well as a tendency towards oversized
data packets, VMC packets should ideally remain local. Thus, in order to
best support the use of VMC with dual-PC setups, this means that the
Sender now needs to be able to receive inputs over OSC, effectively
taking the Viewer's place. Eventually, the Sender will assimilate the
Viewer application and just become the Remote Wheel application, but
it's not there quite yet.

Finally, the Sender's config file format was changed to TOML, rather
than YAML. The use of YAML (and particularly the serde_yaml crate) was a
growing frustration, as it handles flatten and untagged enums very
poorly. Thus, the config file format has been changed to TOML and the
config structure has been reworked to make it more suitable for TOML.