Config structure tweaks and default config update.
This changes the VMC send and receive addresses to match how they are
specified for OSC. The on-change events (for both OSC and VMC) are
renamed to on-update, as they are not verified to be change-only via OSC
or with mixed sources.

OscParameter parsing is now done manually, for more direct control. this
allows ints, floats, strings, and bool values to be specified without a
having to use a table to tag them.