Added choice of default configuration on first start.
This should hopefully help simplify the process of getting started, by
allowing the user to select a default configuration which is focused on
their end-goal. This effectively takes the sample configurations that
were added previously and embeds them into the application, allowing
the user to pick which one to extract when no configuration is present.

This also changes the controller.axis configuration value to take a
number, rather than an axis name. SDL doesn't seem to guarantee any
specific order to the axes, so assigning names to the axis indices was a
bad assumption to begin with. The names are still accepted, for
backwards-compatibility reasons.