Added a Technique subtype to the Wheel device.
This allows the animation for the arms to be encapsulated and switched
out. Currently, there is only one implemented Technique ("Glue") which
is the same as it was before, with the hands fixed to the wheel (though
the position can now be tweaked via the config).

A more advanced Technique is next on the to-do list.